rag-backend/rag-chatbot/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chatbot"
version = "0.1.0"
description = "RAG Pipeline for ingesting website content, generating embeddings, and storing in vector database"
readme = "README.md"
authors = [{name = "AI Developer", email = "developer@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "langchain>=0.0.200",
    "langchain-text-splitters",
    "cohere>=4.0.0",
    "qdrant-client>=1.3.0",
    "python-dotenv>=0.19.0",
    "pytest>=7.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "openai>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950"
]

[project.scripts]
rag-chatbot = "src.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]